exp_rs::eval::stack_ops

Enum EvalOp

pub enum EvalOp<'arena> {

    Eval {
        expr: &'arena AstExpr<'arena>,
        ctx_id: usize,
    },
    ApplyUnary {
        op: UnaryOp,
    },
    CompleteBinary {
        op: BinaryOp,
    },
    ShortCircuitAnd {
        right_expr: &'arena AstExpr<'arena>,
        ctx_id: usize,
    },
    ShortCircuitOr {
        right_expr: &'arena AstExpr<'arena>,
        ctx_id: usize,
    },
    CompleteAnd,
    CompleteOr,
    ApplyFunction {
        name: FunctionName,
        args_needed: usize,
        args_collected: Vec<Real>,
        ctx_id: usize,
    },
    CollectFunctionArgs {
        name: FunctionName,
        total_args: usize,
        args_so_far: Vec<Real>,
        ctx_id: usize,
    },
    TernaryCondition {
        true_branch: &'arena AstExpr<'arena>,
        false_branch: &'arena AstExpr<'arena>,
        ctx_id: usize,
    },
    LookupVariable {
        name: HString,
        ctx_id: usize,
    },
    AccessArray {
        array_name: HString,
        ctx_id: usize,
    },
    AccessAttribute {
        object_name: HString,
        attr_name: HString,
        ctx_id: usize,
    },

}

Operations that can be pushed onto the evaluation stack

Variants

Eval

Push an expression to evaluate

Fields

expr: &'arena AstExpr<'arena>

ctx_id: usize

ApplyUnary

Apply a unary operation

Fields

op: UnaryOp

CompleteBinary

Apply a binary operation after both operands are evaluated

Fields

op: BinaryOp

ShortCircuitAnd

Short-circuit AND operation

Fields

right_expr: &'arena AstExpr<'arena>

ctx_id: usize

ShortCircuitOr

Short-circuit OR operation

Fields

right_expr: &'arena AstExpr<'arena>

ctx_id: usize

CompleteAnd

Complete AND operation (when not short-circuited)

CompleteOr

Complete OR operation (when not short-circuited)

ApplyFunction

Apply a function with N arguments

Fields

name: FunctionName

args_needed: usize

args_collected: Vec<Real>

ctx_id: usize

CollectFunctionArgs

Collect function arguments from the value stack

Fields

name: FunctionName

total_args: usize

args_so_far: Vec<Real>

ctx_id: usize

TernaryCondition

Handle ternary operator - condition already evaluated

Fields

true_branch: &'arena AstExpr<'arena>

false_branch: &'arena AstExpr<'arena>

ctx_id: usize

LookupVariable

Variable lookup

Fields

name: HString

ctx_id: usize

AccessArray

Array access - index already evaluated

Fields

array_name: HString

ctx_id: usize

AccessAttribute

Attribute access

Fields

object_name: HString

attr_name: HString

ctx_id: usize