exp_rs::ffi

Function exp_rs_estimate_arena_size

#[no_mangle]
pub extern "C" fn exp_rs_estimate_arena_size(
    expressions: *const *const c_char,
    num_expressions: usize,
    estimated_iterations: usize,
) -> usize

Estimates the arena size needed for a set of expressions.

This function helps determine the appropriate arena size to allocate for a given set of expressions, preventing frequent reallocations.

Parameters

expressions - Array of expression strings to estimate

num_expressions - Number of expressions in the array

estimated_iterations - Estimated number of evaluation iterations

Returns

Estimated arena size in bytes

Safety

The expressions pointer must be valid and point to at least num_expressions null-terminated C strings.

Example

const char* exprs[] = {"x + y", "sin(x) * cos(y)", "sqrt(x*x + y*y)"};
size_t arena_size = exp_rs_estimate_arena_size(exprs, 3, 1000);
Arena* arena = exp_rs_arena_new(arena_size);